refactor(db): move wcmExperts to repositories/db + services/db
Matches the layout agreed on #832: queries -> repositories/db/,
non-request logic -> services/db/, request/response stays in controllers/db/.

wcmExperts.controller.ts had no req/res handler — findWcmExperts takes
(meshTerms, limit) — so it moved to services/db/wcmExperts.service.ts.
Import updated in src/pages/api/literature/search.ts. No behavior or
query changes.
